Robustify X.509 certificate parsing.
XML with a newline at the beginning of the embedded X.509 cert's data has been encountered in the wild.
This resulted in a "nested asn1 error" when OpenSSL tried to read the data. However, that cert can't be
included in this open-source project, and the behavior couldn't be duplicated with other available certs.