Empty certificate store no longer causes embedded key to be used.
It has always been the intention that if the user provides a certificate
store, validation will *only* be performed using a key in that store
that matches the one in the signature's `KeyInfo` element. This is a
security measure: don't blindly trust that the embedded key is the right
one.

That worked as intended if the provided store had at least one key. If
it was the wrong key, validation failed and a warning message was
printed.

If, however, the provided store had no keys at all, then validation
silently reverted to using the key embedded in the signature.

Fixed that.